Add production deployment for circomspect
This adds a new github actions workflow which builds and deploys optimized
images for every circomspect tag. This runs whenever a PR is delivered, and
also runs nightly at 3 or 4am Eastern. The workflow can also be run manually.
The PR job now deploys all of the images with a `dev` tag instead of `latest`.

Merges #8